requirements/common.txt
Django==4.0.4
django-bootstrap-form==3.4
django-filter==21.1
django-static-precompiler==2.1
django-utils-six==2.0
djangorestframework==3.13.1
openpyxl==3.0.9
Pillow==9.1.0


